[JENKINS-38805_User_can_run_parametrised_pipeline] refactor code to give more structure
