+++ blueocean-dashboard/src/main/js/components/parameter/components/Text.jsx
@@ -1,18 +1,17 @@
 import React, { Component } from 'react';
-import { TextInput, FormElement } from '@jenkins-cd/design-language';
-import { propTypes } from './commonProptypes';
+import { TextArea, FormElement } from '@jenkins-cd/design-language';
+import { propTypes } from '../commonProptypes';
 
-export class String extends Component {
-
+export class Text extends Component {
     render() {
         const { defaultParameterValue: { value }, description, name, onChange } = this.props;
         return (<FormElement title={ name }>
-            <div className="String">
-                <TextInput {...{ defaultValue: value, name, onChange }} />
+            <div className="Text">
+                <TextArea {...{ defaultValue: value, name, onChange }} />
                 { description && <div className="inputDescription">{description}</div> }
             </div>
         </FormElement>);
     }
 }
 
-String.propTypes = propTypes;
+Text.propTypes = propTypes;
