+++ blueocean-dashboard/src/main/js/components/parameter/index.js
@@ -1,10 +1,15 @@
-import { Boolean } from './Boolean';
-import { Choice } from './Choice';
-import { String } from './String';
-import { Text } from './Text';
-import { Password } from './Password';
-export { ParameterService } from './ParameterService';
-export { ParametersRender } from './ParametersRender';
+// Implementations
+import { Boolean } from './components/Boolean';
+import { Choice } from './components/Choice';
+import { String } from './components/String';
+import { Text } from './components/Text';
+import { Password } from './components/Password';
+export { Boolean, Choice, String, Text, Password };
+// Renderer
+export { ParametersRender } from './renderer/ParametersRender';
+export { DebugRender } from './renderer/DebugRender';
+// service and apis
+import { ParameterService } from './services/ParameterService';
 import { ParameterApi } from './rest/ParameterApi';
 
 /**
@@ -27,3 +32,6 @@ export const supportedInputTypes = Object.keys(supportedInputTypesMapping);
 
 const parameterApi = new ParameterApi();
 export { parameterApi as ParameterApi };
+
+const parameterService = new ParameterService();
+export { parameterService as ParameterService};
